move connection service to repository
